+++ src/index.ts
@@ -1,2 +1,3 @@
+export { createExpressApp } from './api/index.js';
 export { validateTokenMiddleware, createAuthApp } from './auth/index.js';
 export { Repository } from './db/index.js';
